fix unit test

+++ service/history/stateBuilder_test.go
@@ -202,10 +202,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeWorkflowExecutionTimedOut()
 	now := time.Now()
 	evenType := shared.EventTypeWorkflowExecutionTimedOut
 	event := &shared.HistoryEvent{
-		Version:                                  common.Int64Ptr(version),
-		EventId:                                  common.Int64Ptr(130),
-		Timestamp:                                common.Int64Ptr(now.UnixNano()),
-		EventType:                                &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		WorkflowExecutionTimedOutEventAttributes: &shared.WorkflowExecutionTimedOutEventAttributes{},
 	}
 
@@ -298,10 +298,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeWorkflowExecutionSignaled()
 	now := time.Now()
 	evenType := shared.EventTypeWorkflowExecutionSignaled
 	event := &shared.HistoryEvent{
-		Version:                                  common.Int64Ptr(version),
-		EventId:                                  common.Int64Ptr(130),
-		Timestamp:                                common.Int64Ptr(now.UnixNano()),
-		EventType:                                &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		WorkflowExecutionSignaledEventAttributes: &shared.WorkflowExecutionSignaledEventAttributes{},
 	}
 	s.mockUpdateVersion(event)
@@ -329,10 +329,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeWorkflowExecutionFailed() {
 	now := time.Now()
 	evenType := shared.EventTypeWorkflowExecutionFailed
 	event := &shared.HistoryEvent{
-		Version:                                common.Int64Ptr(version),
-		EventId:                                common.Int64Ptr(130),
-		Timestamp:                              common.Int64Ptr(now.UnixNano()),
-		EventType:                              &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		WorkflowExecutionFailedEventAttributes: &shared.WorkflowExecutionFailedEventAttributes{},
 	}
 
@@ -477,7 +477,7 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeWorkflowExecutionContinuedA
 			Attempt:         newRunDecisionAttempt,
 		},
 		mock.Anything,
-		0,
+		int32(0),
 	).Return(nil)
 	s.mockUpdateVersion(continueAsNewEvent)
 
@@ -593,10 +593,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeWorkflowExecutionCanceled()
 	now := time.Now()
 	evenType := shared.EventTypeWorkflowExecutionCanceled
 	event := &shared.HistoryEvent{
-		Version:                                  common.Int64Ptr(version),
-		EventId:                                  common.Int64Ptr(130),
-		Timestamp:                                common.Int64Ptr(now.UnixNano()),
-		EventType:                                &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		WorkflowExecutionCanceledEventAttributes: &shared.WorkflowExecutionCanceledEventAttributes{},
 	}
 
@@ -1728,10 +1728,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeActivityTaskStarted() {
 	timeoutSecond := int32(10)
 	evenType := shared.EventTypeActivityTaskScheduled
 	scheduledEvent := &shared.HistoryEvent{
-		Version:                              common.Int64Ptr(version),
-		EventId:                              common.Int64Ptr(130),
-		Timestamp:                            common.Int64Ptr(now.UnixNano()),
-		EventType:                            &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		ActivityTaskScheduledEventAttributes: &shared.ActivityTaskScheduledEventAttributes{},
 	}
 
@@ -1799,10 +1799,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeActivityTaskScheduled() {
 	timeoutSecond := int32(10)
 	evenType := shared.EventTypeActivityTaskScheduled
 	event := &shared.HistoryEvent{
-		Version:                              common.Int64Ptr(version),
-		EventId:                              common.Int64Ptr(130),
-		Timestamp:                            common.Int64Ptr(now.UnixNano()),
-		EventType:                            &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		ActivityTaskScheduledEventAttributes: &shared.ActivityTaskScheduledEventAttributes{},
 	}
 
@@ -1918,10 +1918,10 @@ func (s *stateBuilderSuite) TestApplyEvents_EventTypeActivityTaskCompleted() {
 	now := time.Now()
 	evenType := shared.EventTypeActivityTaskCompleted
 	event := &shared.HistoryEvent{
-		Version:                              common.Int64Ptr(version),
-		EventId:                              common.Int64Ptr(130),
-		Timestamp:                            common.Int64Ptr(now.UnixNano()),
-		EventType:                            &evenType,
+		Version:   common.Int64Ptr(version),
+		EventId:   common.Int64Ptr(130),
+		Timestamp: common.Int64Ptr(now.UnixNano()),
+		EventType: &evenType,
 		ActivityTaskCompletedEventAttributes: &shared.ActivityTaskCompletedEventAttributes{},
 	}
